Refactor Questionnaire component
